requirements.txt
Django==6.0.2
django-tailwind==4.4.2
django-browser-reload==1.21.0
python-dotenv==1.2.1
django-cors-headers==4.9.0
social-auth-app-django==5.7.0
django-sesame==3.2.3
django-ninja==1.5.3

# Two-factor authentication
django-two-factor-auth==1.18.1
django-otp-yubikey==1.1.0

# Additional dependencies for Turnstile CAPTCHA
requests==2.32.5

# Database adapter for PostgreSQL
psycopg2-binary==2.9.11

# Redis client for caching
redis==7.1.1

# Allows for easy management of package updates
pip-review==1.3.0

# Customizable admin dashboard
django-unfold==0.79.0
